+++ android/app/build.gradle
@@ -190,6 +190,7 @@ android {
 }
 
 dependencies {
+    implementation project(':react-native-blob-util')
     implementation project(':react-native-file-viewer')
     implementation project(':react-native-sound-player')
     implementation project(':react-native-randombytes')
