+++ android/app/src/main/java/lab/childmindinstitute/data/MainApplication.java
@@ -4,6 +4,7 @@ import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
+import com.ReactNativeBlobUtil.ReactNativeBlobUtilPackage;
 import com.vinzscam.reactnativefileviewer.RNFileViewerPackage;
 import com.johnsonsu.rnsoundplayer.RNSoundPlayerPackage;
 import com.bitgo.randombytes.RandomBytesPackage;
