+++ app/services/api.js
@@ -8,7 +8,7 @@ import {
   postFile,
   downloadTokenResponses
 } from "./network";
-import { getData } from "./asyncStorage";
+import { getData } from "./storage";
 import { cleanFiles } from "./file";
 import { transformResponses, decryptAppletResponses } from "../models/response";
 import { decryptData } from "./encryption";
