+++ package.json
@@ -78,6 +78,7 @@
     "react-i18next": "^11.8.4",
     "react-native": "^0.61.5",
     "react-native-awesome-alerts": "^1.2.0",
+    "react-native-blob-util": "^0.13.18",
     "react-native-calendar-strip": "^1.3.8",
     "react-native-camera": "^3.16.0",
     "react-native-crypto": "^2.2.0",
@@ -126,6 +127,7 @@
     "redux-form": "^8.2.6",
     "redux-logger": "^3.0.6",
     "redux-persist": "^5.10.0",
+    "redux-persist-filesystem-storage": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0",
     "rn-fetch-blob": "^0.12.0",
